Add files via upload
Project assigned from Automate the Boring Stuff. The purpose of this project was to write a program that opens all .txt files in a folder and searches within the .txt files for any matches to the user-supplied regular expression. Results are then printed to the screen. 

Any comments or suggestions on how to make this code more efficient are welcome.