Issues with Renaming Files after a Regex Search
Project from "Automate the Boring Stuff with Python" Chapter 9. Exercise is to:  "Write a program that finds all files with a given prefix, such as spam001.txt, spam002.txt, and so on, in a single folder and locates any gaps in the numbering (such as if there is a spam001.txt and spam003.txt but no spam002.txt). Have the program rename all the later files to close this gap."

The program worked fine up until I added code at lines 40 and after. I am trying to rename the files based on whether a gap occurs. In my dummy folder on my computer, I have a folder with sample files such as "eggs001.txt, eggs002.txt, eggs003.txt, notNumberedFile.txt, eggs005.txt" so I know that there is a gap between eggs003.txt and eggs004.txt. I'm not sure if the way I am renaming files is wrong or if there is a better way of doing this.